+++ src/components/netpol/index.test.ts
@@ -1,11 +1,19 @@
-//
+import environmentMock from "@socialgouv/kosko-charts/environments/index.mock";
 
 beforeEach(() => {
   jest.resetModules();
 });
 
-test("should create a network policy", async () => {
+test("should create a network policy with custom namespace", async () => {
+  jest.doMock("@socialgouv/kosko-charts/environments", () => environmentMock);
   const { create } = await import("./index");
   const netpol = create("some-app");
+  expect(netpol).toMatchSnapshot();
+});
+
+test("should create a network policy", async () => {
+  jest.doMock("@socialgouv/kosko-charts/environments", () => environmentMock);
+  const { create } = await import("./index");
+  const netpol = create();
   expect(netpol).toMatchSnapshot();
 });