+++ src/components/netpol/index.ts
@@ -1,3 +1,4 @@
+import environments from "@socialgouv/kosko-charts/environments";
 import { NetworkPolicy } from "kubernetes-models/networking.k8s.io/v1/NetworkPolicy";
 
 /**  Basic netpol that allow :
@@ -6,17 +7,17 @@ import { NetworkPolicy } from "kubernetes-models/networking.k8s.io/v1/NetworkPol
 /*      - communication from monitoring
 */
 
-export const create = (namespace: string): NetworkPolicy =>
-  new NetworkPolicy({
-    metadata: { name: "default", namespace },
+export const create = (namespace?: string): NetworkPolicy => {
+  const ciEnv = environments(process.env);
+  const ns = namespace ?? ciEnv.metadata.namespace.name;
+  return new NetworkPolicy({
+    metadata: { name: `netpol-${ns}`, namespace: ns },
     spec: {
       ingress: [
         {
           from: [
             {
-              namespaceSelector: {
-                matchLabels: { "network-policy/namespace": namespace },
-              },
+              podSelector: {},
             },
           ],
         },
@@ -47,3 +48,4 @@ export const create = (namespace: string): NetworkPolicy =>
       policyTypes: ["Ingress"],
     },
   });
+};